requirements.txt
python-dotenv
streamlit
pandasai==3.0.0b16
pandas
pandasai-openai
numpy==1.26.4
xlrd
pandasql
pycryptodomex